fix(ci): repair main-merge fallout — LayerPool k/v strides + build.rs size cap
The merge of main left two semantic conflicts:

1. spark-runtime E0609 x6: this branch splits LayerPool.block_stride into
   k_block_stride/v_block_stride (asymmetric K/V dtypes), but main-side
   diagnostics (poison_block, the BF16 KV checksum/per-block probes) still
   referenced the old single field. poison_block now strides K and V
   independently; the two BF16-only probes use k_block_stride (== v for
   symmetric dtypes, and they early-return unless dtype is Bf16).

2. factory/build.rs hit 505 LoC (cap 500) after absorbing main's
   --lm-head-dtype wiring. Extracted the Step-3 LM-head quantization block
   (NVFP4 / FP8 / BF16-skip selection + the draft-only NVFP4 MTP head)
   into factory/lm_head_setup.rs as setup_lm_heads() — pure code move,
   build.rs down to 411.

cargo check/clippy/fmt clean; cargo test --workspace 1554 passed 0 failed.
